test: separate into Portal.test.js and PortalManager.test.js
